Update copyrights, move modules to better names, etc etc
