+++ base64.cabal
@@ -26,7 +26,7 @@ source-repository head
 
 library
   exposed-modules:
-    Data.Base64
+    Data.Base64.Types
     Data.ByteString.Base64
     Data.ByteString.Base64.URL
     Data.ByteString.Lazy.Base64
@@ -42,7 +42,7 @@ library
     Data.Text.Short.Encoding.Base64.URL
 
   other-modules:
-    Data.Base64.Internal
+    Data.Base64.Types.Internal
     Data.ByteString.Base64.Internal
     Data.ByteString.Base64.Internal.Head
     Data.ByteString.Base64.Internal.Tables
