+++ src/Data/ByteString/Base64.hs
@@ -5,7 +5,7 @@
 {-# LANGUAGE Trustworthy #-}
 -- |
 -- Module       : Data.ByteString.Base64
--- Copyright    : (c) 2019-2020 Emily Pillmore
+-- Copyright    : (c) 2019-2022 Emily Pillmore
 -- License      : BSD-style
 --
 -- Maintainer   : Emily Pillmore <emilypi@cohomolo.gy>
@@ -29,8 +29,8 @@ module Data.ByteString.Base64
 , isValidBase64
 ) where
 
-import Data.Base64
-import Data.Base64.Internal
+import Data.Base64.Types
+import Data.Base64.Types.Internal
 
 import Data.ByteString.Internal (ByteString(..))
 import Data.ByteString.Base64.Internal
