+++ src/Data/ByteString/Lazy/Base64/URL.hs
@@ -5,7 +5,7 @@
 {-# LANGUAGE Trustworthy #-}
 -- |
 -- Module       : Data.ByteString.Lazy.Base64.URL
--- Copyright    : (c) 2019-2020 Emily Pillmore
+-- Copyright    : (c) 2019-2022 Emily Pillmore
 -- License      : BSD-style
 --
 -- Maintainer   : Emily Pillmore <emilypi@cohomolo.gy>
@@ -36,7 +36,7 @@ module Data.ByteString.Lazy.Base64.URL
 
 import Prelude hiding (all, elem)
 
-import Data.Base64
+import Data.Base64.Types
 
 import qualified Data.ByteString as BS
 import qualified Data.ByteString.Base64.URL as B64U
