+++ src/Data/Text/Lazy/Encoding/Base64.hs
@@ -2,7 +2,7 @@
 {-# LANGUAGE Safe #-}
 -- |
 -- Module       : Data.Text.Lazy.Encoding.Base64
--- Copyright    : (c) 2019-2020 Emily Pillmore
+-- Copyright    : (c) 2019-2022 Emily Pillmore
 -- License      : BSD-style
 --
 -- Maintainer   : Emily Pillmore <emilypi@cohomolo.gy>
@@ -27,7 +27,7 @@ module Data.Text.Lazy.Encoding.Base64
 ) where
 
 
-import Data.Base64
+import Data.Base64.Types
 
 import Data.Bifunctor (first)
 import Data.ByteString.Lazy (ByteString)
