+++ src/Data/Text/Short/Encoding/Base64.hs
@@ -2,7 +2,7 @@
 {-# LANGUAGE Trustworthy #-}
 -- |
 -- Module       : Data.Text.Short.Encoding.Base64
--- Copyright    : (c) 2019-2020 Emily Pillmore
+-- Copyright    : (c) 2019-2022 Emily Pillmore
 -- License      : BSD-style
 --
 -- Maintainer   : Emily Pillmore <emilypi@cohomolo.gy>
@@ -26,7 +26,7 @@ module Data.Text.Short.Encoding.Base64
 , isValidBase64
 ) where
 
-import Data.Base64
+import Data.Base64.Types
 
 import Data.Bifunctor (first)
 import qualified Data.ByteString.Base64 as B64
